Create modules under neksuite
